refactor(app): 抽取 composer 状态模型到独立文件
